Fixed a sign-comparison warning
Fixes #35

+++ Source/TDMultiInputStream.m
@@ -236,7 +236,7 @@ TestCase(TDConcatenatedInputStream_Sync) {
         uint8_t buffer[bufSize];
         NSInteger nBytes;
         while ((nBytes = [mp read: buffer maxLength: sizeof(buffer)]) > 0) {
-            CAssert(nBytes <= bufSize);
+            CAssert((unsigned)nBytes <= bufSize);
             [output appendBytes: buffer length: nBytes];
         }
         CAssert(nBytes == 0, @"Stream returned an error");
